+++ src/main/java/com/arturo/releaseradarapi/repository/TituloRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface TituloRepository extends JpaRepository<Titulo, Long> {
-    Optional<Titulo> findByTmdbId(Integer tmdbId);
+    Optional<Titulo> findByTmdbId(Long tmdbId);
 }
